Add files via upload
I made a simple header with a temporary cream color that we can definitely discuss if we want to keep but I thought I would try a new color. Buttons can be added in once we have the code for them although it may have been easier to just code the buttons into the header in the first place.